Moved formatter into namespace
